move
토큰, 레디스 관련 클래스 이동
